Fixes #27
